More compact and readable files structure, minor improvments.
