--- backend/src/routers/login.py
@@ -1,162 +1,0 @@
-from datetime import datetime, timedelta
-import jwt
-from fastapi import Depends, HTTPException, APIRouter
-from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
-from jwt import PyJWTError
-from pydantic import BaseModel
-from starlette.status import HTTP_401_UNAUTHORIZED
-import json
-from starlette.responses import Response
-from fastapi.encoders import jsonable_encoder
-from passlib.context import CryptContext
-
-# to get a string like this run:
-# openssl rand -hex 32
-
-# Secret key will be moved from here, development purposes only.
-SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
-
-ALGORITHM = "HS256"
-REFRESH_TOKEN_EXPIRE_MINUTES = 360
-ACCES_TOKEN_EXPIRES_MINUTES = 15
-
-MOCK_DB = {
-    "johndoe": {
-        "username": "ada",
-        "full_name": "John Doe",
-        "email": "johndoe@example.com",
-        "hashed_password": "$2b$12$EixZaYVK1fsbw1ZfbX3OX\
-ePaWxn96p36WQoeG6Lruj3vjPGga31lW",
-        "disabled": False,
-    }
-}
-
-PWD_CONTEXT = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-OAUTH2_SCHEME = OAuth2PasswordBearer(tokenUrl="/login")
-
-router = APIRouter()
-
-
-class Token(BaseModel):
-    access_token: str
-    token_type: str
-
-
-class TokenData(BaseModel):
-    username: str = None
-
-
-class User(BaseModel):
-    username: str
-    email: str = None
-    full_name: str = None
-    disabled: bool = None
-
-
-class UserInDB(User):
-    hashed_password: str
-
-
-def verify_password(plain_password, hashed_password):
-    return PWD_CONTEXT.verify(plain_password, hashed_password)
-
-
-def get_password_hash(password):
-    return PWD_CONTEXT.hash(password)
-
-
-def get_user(db, username: str):
-    if username in db:
-        user_dict = db[username]
-        return UserInDB(**user_dict)
-    return None
-
-
-def authenticate_user(fake_db, username: str, password: str):
-    user = get_user(fake_db, username)
-    if not user:
-        return False
-    if not verify_password(password, user.hashed_password):
-        return False
-    return user
-
-
-def create_token(*, data: dict, expires_delta: timedelta = None):
-    to_encode = data.copy()
-    if expires_delta:
-        expire = datetime.utcnow() + expires_delta
-    else:
-        expire = datetime.utcnow() + timedelta(minutes=15)
-    to_encode.update({"exp": expire})
-    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-    return encoded_jwt
-
-async def get_current_user(token: str = Depends(OAUTH2_SCHEME)):
-    credentials_exception = HTTPException(
-        status_code=HTTP_401_UNAUTHORIZED,
-        detail="Could not validate credentials",
-        headers={"WWW-Authenticate": "Bearer"},
-    )
-    try:
-        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
-        username: str = payload.get("sub")
-        if username is None:
-            raise credentials_exception
-        token_data = TokenData(username=username)
-    except PyJWTError:
-        raise credentials_exception
-    user = get_user(MOCK_DB, username=token_data.username)
-    if user is None:
-        raise credentials_exception
-    return user
-
-
-async def get_current_active_user(current_user: User = Depends(get_current_user)):
-    if current_user.disabled:
-        raise HTTPException(status_code=400, detail="Inactive user")
-    return current_user
-
-
-@router.post("/login", response_model=Token)
-async def login_for_access_token(form_data: OAuth2PasswordRequestForm = Depends()):
-    user = authenticate_user(MOCK_DB, form_data.username, form_data.password)
-    if not user:
-        print("User NOT verified.")
-        raise HTTPException(
-            status_code=HTTP_401_UNAUTHORIZED,
-            detail="Incorrect username or password",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-    
-    refresh_token_expires = timedelta(minutes=REFRESH_TOKEN_EXPIRE_MINUTES)
-    access_token_expires = timedelta(minutes=ACCES_TOKEN_EXPIRES_MINUTES)
-
-
-    refresh_token = create_token(
-        data={"sub": form_data.username}, expires_delta=refresh_token_expires)
-    acces_token = create_token(
-        data={"sub":form_data.username}, expires_delta=access_token_expires)
-    ref_token = jsonable_encoder(refresh_token)
-    acc_token = jsonable_encoder(acces_token)
-
-    print("User succesfully verified.")
-    response = Response(status_code=200)
-    
-    response.set_cookie(
-        key="refresh_token",
-        value=f"{ref_token}",
-        httponly=True,
-    )
-    response.set_cookie(
-        key="access_token",
-        value=f"{acc_token}",
-        httponly=True,   
-    )
-
-    return response
-
-
-@router.get("/users/me/", response_model=User)
-async def read_users_me(current_user: User = Depends(get_current_active_user)):
-    return current_user
